printing ticks also
git-svn-id: https://trigonak-thesis.googlecode.com/svn/exjobb/iRCCEdeddsl@4027 dd814760-9b98-ad33-1d6a-b699670a4318

+++ include/measurements.h
@@ -222,7 +222,7 @@ while (0);
                         ticks_perc, total_samples[i],
                         s, ms, us, ns,
                         sa, msa, usa, nsa,
-                        total_sum_ticks[i]);
+                        (long long int) total_sum_ticks[i]);
             }
         }
